+++ src/device/humidifier.ts
@@ -662,7 +662,7 @@ export class Humidifier {
   }
 
   async stopScanning(switchbot: any) {
-    switchbot.stopScan();
+    await switchbot.stopScan();
     if (this.connected) {
       await this.BLEparseStatus();
       await this.updateHomeKitCharacteristics();
@@ -684,7 +684,7 @@ export class Humidifier {
         };
         await sleep(10000);
         // Stop to monitor
-        switchbot.stopScan();
+        await switchbot.stopScan();
       })();
     }
   }