+++ src/device/robotvacuumcleaner.ts
@@ -523,7 +523,7 @@ export class RobotVacuumCleaner {
   }
 
   async stopScanning(switchbot: any) {
-    switchbot.stopScan();
+    await switchbot.stopScan();
     if (this.connected) {
       await this.BLEparseStatus();
       await this.updateHomeKitCharacteristics();
@@ -553,7 +553,7 @@ export class RobotVacuumCleaner {
         };
         await sleep(10000);
         // Stop to monitor
-        switchbot.stopScan();
+        await switchbot.stopScan();
       })();
     }
   }