+++ src/platform.ts
@@ -23,11 +23,12 @@ import { AirPurifier } from './irdevice/airpurifier';
 import { WaterHeater } from './irdevice/waterheater';
 import { VacuumCleaner } from './irdevice/vacuumcleaner';
 import { AirConditioner } from './irdevice/airconditioner';
-import {request} from 'undici';
+import { request } from 'undici';
 import crypto, { randomUUID } from 'crypto';
 import { Buffer } from 'buffer';
 import { queueScheduler } from 'rxjs';
 import fakegato from 'fakegato-history';
+import { EveHomeKitTypes } from 'homebridge-lib';
 
 import { readFileSync, writeFileSync } from 'fs';
 import { API, DynamicPlatformPlugin, Logger, PlatformAccessory, Service, Characteristic } from 'homebridge';
@@ -51,6 +52,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
   platformLogging?: string;
 
   public readonly fakegatoAPI: any;
+  public readonly eve: any;
 
   constructor(public readonly log: Logger, public readonly config: SwitchBotPlatformConfig, public readonly api: API) {
     this.logs();
@@ -76,8 +78,9 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       return;
     }
 
-    // import fakegato-history module
+    // import fakegato-history module and EVE characteristics
     this.fakegatoAPI = fakegato(api);
+    this.eve = new EveHomeKitTypes(api);
 
     // When this event is fired it means Homebridge has restored all cached accessories from disk.
     // Dynamic Platform plugins should only register new accessories after this event was fired,
@@ -292,13 +295,17 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         const deviceLists = devicesAPI.body.deviceList;
         if (!this.config.options?.devices) {
           this.debugLog(`SwitchBot Device Config Not Set: ${JSON.stringify(this.config.options?.devices)}`);
-          const devices = deviceLists.map((v: any) => v);
-          for (const device of devices) {
-            if (device.deviceType) {
-              if (device.configDeviceName) {
-                device.deviceName = device.configDeviceName;
+          if (devicesAPI.body.length === 0) {
+            this.debugLog(`SwitchBot API Currently Doesn't Have Any Devices With Cloud Services Enabled: ${JSON.stringify(devicesAPI.body)}`);
+          } else {
+            const devices = deviceLists.map((v: any) => v);
+            for (const device of devices) {
+              if (device.deviceType) {
+                if (device.configDeviceName) {
+                  device.deviceName = device.configDeviceName;
+                }
+                this.createDevice(device);
               }
-              this.createDevice(device);
             }
           }
         } else if (this.config.credentials?.token && this.config.options.devices) {
@@ -1635,13 +1642,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     }
   }
 
-  private async createAirConditioner(device: irdevice & devicesConfig) {
+  private async createAirConditioner(device: irdevice & devicesConfig & irDevicesConfig) {
     const uuid = this.api.hap.uuid.generate(`${device.deviceId}-${device.remoteType}`);
 
     // see if an accessory with the same uuid has already been registered and restored from
     // the cached devices we stored in the `configureAccessory` method above
     const existingAccessory = this.accessories.find((accessory) => accessory.UUID === uuid);
 
+    if (device.irair?.meterType && device.irair?.meterId) {
+      device.irair.meterUuid = this.api.hap.uuid.generate(`${device.irair.meterId}-${device.irair.meterType}`);
+    }
     if (existingAccessory) {
       // the accessory already exists
       if (!device.hide_device && device.hubDeviceId) {